\name{basicpayoffs}
\alias{basicpayoffs}
\title{
Plot payoffs / profit and loss of European Call/Put.
}
\description{
Function to Plot payoffs / profit and loss of European Call/Put.
}
\usage{
basicpayoffs()
}
\details{
The user inputs are as follows:\cr
Position:  choose Long/Short/both\cr
Option Type:  chosen between Call/Put\cr
Plot Type:  chosen between Payoff/Profit-and-Loss\cr
}


\value{
Plot of payoffs / profit and loss of European Call/Put.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{trading.menu}}
}

