\name{bearspreadputs}
\alias{bearspreadputs}
\title{
Profit & Loss plot of bear spread with puts.
}
\description{
Function to plot Profit & Loss of bear spread with puts.
}
\usage{
bearspreadputs()
}
\details{
Short put check box: checking it plots the Profit and loss of a short put position.\cr
Long put check box: checking it plots the Profit and loss of a long put position.\cr
Profit check box: checking it plots the over all Profit and loss of a bear spread with puts.
}


\value{
Profit & Loss plot of bear spread with puts.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{bullspreadcalls},\link[GUIDE]{trading.menu}}
}


