\name{BrownianPaths}
\alias{BrownianPaths}
\title{
Simulate and plot Brownian Motion path(s)
}
\description{
Function to simulate and plot Brownian Motion path(s)

}
\usage{
BrownianPaths()
}
\details{
The user inputs are as follows:\cr
Paths\cr
Clicking on the '+' and '-' respectively increases and decreases the values of each of the input.
}

\value{
A graph of Brownian Motion path(s).
}
\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{GBMPaths},\link[GUIDE]{ABMPaths}}
}
