\name{bullspreadcalls}
\alias{bullspreadcalls}
\title{
Profit & Loss plot of bull spread with calls.
}
\description{
Function to plot Profit & Loss of bull spread with calls.
}
\usage{
bullspreadcalls()
}
\details{
Long call check box: checking it plots the Profit and loss of a long call position.\cr
Short call check box: checking it plots the Profit and loss of a short call position.\cr
Profit check box: checking it plots the over all Profit and loss of a bull spread with calls.
}


\value{
Profit & Loss plot of bull spread with calls.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{bearspreadputs},\link[GUIDE]{trading.menu}}
}

