\name{butterfly}
\alias{butterfly}
\title{
Profit & Loss plot of butterfly.
}
\description{
Function to plot Profit & Loss of butterfly.
}
\usage{
butterfly()
}
\details{
Long call 1 check box: checking it plots the Profit and loss of a long call position.\cr
Long call 2 check box: checking it plots the Profit and loss of a long call position.\cr
Short two calls check box: checking it plots the Profit and loss of 2 short calls position.\cr
Profit check box: checking it plots the over all Profit and loss of a butterfly.
}


\value{
Profit & Loss plot of butterfly.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{reversebutterfly},\link[GUIDE]{trading.menu}}
}

