\name{cashprice}
\alias{cashprice}
\title{
Calculate the Cash price of a T Bond Futures
}
\description{
Function to calculate the Cash price of a T Bond Futures

}
\usage{
cashprice()
}
\details{
The user inputs are as follows:\cr
Quoted Price: e.g. 97.8\cr
Conv. Factor: e.g. 1.06\cr
Acc. Interest: in dollars e.g. 3.50\cr
}


\value{
The Cash price of a T Bond Futures.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{futurescurrency},\link[GUIDE]{futurescommodity}}
}
