% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_mock_dir}
\alias{create_mock_dir}
\title{Create a Mock Directory Structure}
\usage{
create_mock_dir(dir_structure, .env = parent.frame())
}
\arguments{
\item{dir_structure}{A named list defining the directory and file structure.
Directories are represented by lists, while files are represented by `TRUE`.}

\item{.env}{The environment in which to schedule the cleanup. Default is
`parent.frame()`.}
}
\value{
A list containing:
  \item{path}{The path to the created mock directory.}
  \item{here}{A function that mimics `here::here()` to access files within
    the mock directory.}
}
\description{
This function creates a temporary directory structure based on the provided
directory structure. It sets up directories and files as specified and
provides a function to mimic `here::here()` behavior for accessing files
within the mock directory. The temporary directory is automatically cleaned up
after use.
}
\details{
The function creates a temporary directory at a location determined by
`tempdir()`, sets up the specified directory and file structure, and returns
a list containing the path to the mock directory and a function to access
files within it. The working directory is temporarily changed to the mock
directory during the operation, and the original working directory is restored
after cleanup.
}
\keyword{internal}
