\name{durcoupon}
\alias{durcoupon}
\title{
Plot the relationship between duration and coupon rate of a bond.
}
\description{
Function to Plot the relationship between duration and coupon rate of a bond.
}
\usage{
durcoupon()
}
\details{
The user inputs are as follows:\cr
Discount Rate (%  p.a)\cr
Maturity (Yrs)\cr
Clicking on "+/-" increases/decreases the values of the above two inputs
}


\value{
A Plot of the reltionship between duration and coupon rate of a bond.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{bondchange},\link[GUIDE]{bondprice}}
}
