\name{eurodollar}
\alias{eurodollar}
\title{
Calculate the value of a eurodollar futures contract price from the CME IMM Quote.
}
\description{
Function to calculate value of a eurodollar futures contract price for notional of 1 mn from the CME IMM Quote.

}
\usage{
eurodollar()
}
\details{
The user inputs are as follows:\cr
CME Quote: e.g. 97.8\cr
}


\value{
The value of a eurodollar futures contract price for notional of 1 mn from the CME IMM Quote.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{futurescurrency},\link[GUIDE]{futurescommodity}}
}
