\name{GBMPaths}
\alias{GBMPaths}
\title{
Simulate and plot Geometric Brownian Motion path(s)
}
\description{
Function to simulate and plot Geometric Brownian Motion path(s)
}
\usage{
GBMPaths()
}
\details{
The user inputs are as follows:\cr
Drift (or mu) \cr
Volatility(or sigma) \cr
Paths\cr
Clicking on the '+' and '-' respectively increases and decreases the values of each of the above three inputs.
}

\value{
A plot of Geometric Brownian Motion path(s) showing the Drift rate (mu) and the Volatility (sigma).

}
\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}
\seealso{
\code{\link[GUIDE]{ABMPaths},\link[GUIDE]{BrownianPaths}} 
}
