\name{GUIDE}
\alias{GUIDE}
\title{
The main menu for the GUIDE package.
}
\description{
Function to bring up the main menu for the GUIDE package
}
\usage{
GUIDE()
}
\details{
Entering "GUIDE()" brings up a GUI containing the following menus:\cr
Forwards\cr
Futures\cr
Options\cr
Swaps\cr
Stochastic Processes\cr
Value at Risk\cr
Bonds and\cr
Utilities
}


\value{
The main menu for the GUIDE package
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}


