% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/child.R
\name{guide_to_child}
\alias{guide_to_child}
\title{Guide to Move Working Directory to Child Directory}
\usage{
guide_to_child(child_dir, create = FALSE)
}
\arguments{
\item{child_dir}{A character string specifying the name of the child directory to move into.}

\item{create}{A logical value indicating whether to create the child directory if it does not exist. Defaults to `FALSE`.}
}
\value{
The new working directory (invisibly) after moving into the child directory.
}
\description{
This function moves the working directory to a specified child directory. If the directory does not exist,
the user has the option to create it.
}
\details{
The function attempts to move the working directory to the specified child directory. If the directory does not exist
and `create` is set to `TRUE`, the directory will be created. If `create` is set to `FALSE`, an error will be thrown
if the directory does not exist.
}
\examples{
\dontrun{
# Move to a child directory named "project"
guides::guide_to_child("project")

# Move to a child directory and create it if it doesn't exist
guides::guide_to_child("project", create = TRUE)
}

}
