\name{JDPaths}
\alias{JDPaths}
\title{
Simulate and plot Jump Diffusion path(s)
}
\description{
Function to simulate and plot Jump Diffusion path(s)

}
\usage{
JDPaths()
}
\details{
The user inputs are as follows:\cr
Drift (or mu) \cr
Volatility(or sigma) \cr
Mean of jumps \cr
Std Dev of Jumps \cr
Jump Intensity \cr
Paths\cr
Clicking on the '+' and '-' respectively increases and decreases the values of each of the above inputs.
}

\value{
A graph of Jump Diffusion path(s) showing the parameter values.
}
\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{GBMPaths},\link[GUIDE]{ABMPaths}}
}
