\name{Premium3D}
\alias{Premium3D}
\title{
Option premium as a function of stock price/strike and time.
}
\description{
Function to plot the option premium as a function of stock price/strike and time


}
\usage{
Premium3D()
}
\details{
The user inputs are as follows:\cr
Type of Option:  chosen between Call/Put\cr
X-Y axis:  chosen between Stock price-Time/Strike - Time\cr
sigma\cr
Risk free rate\cr
Clicking on the '+'/'-' respectively increases/decreases the values.}


\value{
A plot of the option premium as a function of stock price/strike and time

}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{basicpayoffs}}
}

