\name{pricematurity}
\alias{pricematurity}
\title{
Plot the relationship between price and maturity of a bond.
}
\description{
Function to Plot the relationship between price and maturity of a bond.
}
\usage{
pricematurity()
}
\details{
The user inputs are as follows:\cr
Coupon rate (per cent  p.a.)\cr
Discount rate (or yield) p.a.\cr
Moving the slider increases/decreases the values of the above two inputs.\cr
Coupon frequency: chosen amongst quarterly, semi-annual and annual.
}


\value{
A Plot of the reltionship between price and maturity of a bond.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{priceyield},\link[GUIDE]{bondprice}}
}
