\name{priceyield}
\alias{priceyield}
\title{
Plot the relationship between price and yield of a bond.
}
\description{
Function to Plot the relationship between price and yield of a bond.
}
\usage{
priceyield()
}
\details{
The user inputs are as follows:\cr
Coupon rate (per cent  p.a.)\cr
Maturity (yrs)\cr
Moving the slider increases/decreases the values of the above two inputs.\cr
Coupon frequency: chosen amongst quarterly, semi-annual and annual.
}


\value{
A Plot of the reltionship between price and yield of a bond.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{pricematurity},\link[GUIDE]{bondprice}}
}
