\name{pv}
\alias{pv}
\title{
Calculate the Present value of an amount.
}
\description{
Function to calculate the Present value of an amount.
}
\usage{
pv()
}
\details{
The user inputs are as follows:\cr
Future Value:  entered in decimals. For e.g. 105.50\cr
Rate:  entered in decimals. For e.g. 5\ per cent is entered as 0.05\cr
Time:  entered in number of years. For e.g. half year is 0.5\cr
Compounding frequency:  chosen amongst continuous/Quarterly/Semi-annual/Annual\cr
}


\value{
The Present value of an amount.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{pvann},\link[GUIDE]{fv}}
}
