\name{rate}
\alias{rate}
\title{
Calculate rate in the desired frequency.
}
\description{
Function to calculate rate in the desired frequency.
}
\usage{
rate()
}
\details{
The user inputs are as follows:\cr
Given frequency:  chosen amongst continuous/Quarterly/Semi-annual/Annual\cr
Required frequency:  chosen amongst continuous/Quarterly/Semi-annual/Annual\cr
Given rate:  entered in decimals. For e.g. 5 per cent is entered as 0.05\cr
}


\value{
Rate expressed the desired frequency.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{pv},\link[GUIDE]{fv}}
}
