\name{ratetreegui}
\alias{ratetreegui}
\title{
Plot a interest rate  tree
}
\description{
Function to plot a interest rate  tree
}
\usage{
ratetreegui()
}
\details{
The user inputs are as follows:\cr
Rate: to be entered in percent. For e.g. enter 5.0 for 5 percent \cr
u: up move factor- to be entered in decimals. For e.g. 1.25 \cr
d: down move factor- to be entered in decimals. For e.g. 0.80\cr
q: probability of up move- to be entered in decimals. For e.g. 0.60\cr
Rate tree steps: Clicking on the '+' and '-' respectively increases and decreases the value.\cr
}


\value{
A plot of interest rate  tree with user specified parameters.
}

\references{
David G. Luenberger, "Investment Science", 2/E, Oxford University Press, 2013. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{bondtreegui}}
}

