\name{reversebutterfly}
\alias{reversebutterfly}
\title{
Profit & Loss plot of reverse butterfly.
}
\description{
Function to plot Profit & Loss of reverse butterfly.
}
\usage{
reversebutterfly()
}
\details{
Short call 1 check box: checking it plots the Profit and loss of a Short call position.\cr
Short call 2 check box: checking it plots the Profit and loss of a Short call position.\cr
Long two calls check box: checking it plots the Profit and loss of 2 long calls position.\cr
Profit check box: checking it plots the over all Profit and loss of a reverse butterfly.
}


\value{
Profit & Loss plot of reverse butterfly.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{butterfly},\link[GUIDE]{trading.menu}}
}

