\name{reversestraddle}
\alias{reversestraddle}
\title{
Profit & Loss plot of reverse straddle.
}
\description{
Function to plot Profit & Loss of reversestraddle.
}
\usage{
reversestraddle()
}
\details{
Long Put check box: checking it plots the Profit and loss of a long put position.\cr
Long Call check box: checking it plots the Profit and loss of a long call position.\cr
Profit check box: checking it plots the over all Profit and loss of a reverse straddle.
}


\value{
Profit & Loss plot of reversestraddle.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{straddle},\link[GUIDE]{trading.menu}}
}

