\name{strap}
\alias{strap}
\title{
Profit & Loss plot of strap.
}
\description{
Function to plot Profit & Loss of strap.
}
\usage{
strap()
}
\details{
Two Long Calls check box: checking it plots the Profit and loss of a long put position.\cr
Long Put check box: checking it plots the Profit and loss of a long call position.\cr
Profit check box: checking it plots the over all Profit and loss of a strap.
}


\value{
Profit & Loss plot of strap.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{strip},\link[GUIDE]{trading.menu}}
}

