\name{strip}
\alias{strip}
\title{
Profit & Loss plot of strip.
}
\description{
Function to plot Profit & Loss of strip.
}
\usage{
strip()
}
\details{
Two Long Puts check box: checking it plots the Profit and loss of a long put position.\cr
Long Call check box: checking it plots the Profit and loss of a long call position.\cr
Profit check box: checking it plots the over all Profit and loss of a strip.
}


\value{
Profit & Loss plot of strip.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{strap},\link[GUIDE]{trading.menu}}
}

