\name{trading.menu}
\alias{trading.menu}
\title{
A menu for Option trading strategies.
}
\description{
Function to bring up the menu for option trading strategies.
}
\usage{
trading.menu()
}
\details{
Brings up a menu of option trading strategies to choose from. Strategies include:\cr
Bull spread\cr
Bear spread\cr
Butterfly\cr
Reverse butterfly\cr
Straddle\cr
Reverse straddle\cr
Strangle\cr
Reverse Strangle\cr
Strip\cr
Strap\cr
\cr
Making a choice plots the chosen trading strategy.}


\value{
A menu of various Option trading strategies.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012.
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{basicpayoffs}}
}
