\name{zval}
\alias{zval}
\title{
Calculate the cumulative probability corresponding to a given a z value from a normal distribution.
}
\description{
Function to calculate the cumulative probability corresponding to a given a z value from a normal distribution.
}
\usage{
zval()
}
\details{
The user input is as follows:\cr
z value: A number that can be from - infinity to + infinity. E.g. -1.65\cr
}


\value{
The cumulative probability from the left tail of the distribution till the given z value.
}

\references{
John C. Hull, "Options, Futures, and Other Derivatives", 8/E, Prentice Hall, 2012. 
}
\author{
S Subramanian <ssubramanian@sssihl.edu.in>
}

\seealso{
\code{\link[GUIDE]{pval}}
}

